backend/requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
djangorestframework==3.10.3
docker==4.0.2
requests==2.22.0